+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 - [feat: support lossless config from console & support warmup.](https://github.com/Tencent/spring-cloud-tencent/pull/1445)
 - [feat:add admin http handler.](https://github.com/Tencent/spring-cloud-tencent/pull/1450)
 - [feat:upgrade spring cloud 2023 version.](https://github.com/Tencent/spring-cloud-tencent/pull/1451)
+- [feat:support concurrency rate limit.](https://github.com/Tencent/spring-cloud-tencent/pull/1455)